% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aopCytoscape_Class.R
\docType{class}
\name{aop_cytoscape-class}
\alias{aop_cytoscape}
\alias{aop_cytoscape-class}
\title{aop_cytoscape class}
\description{
Creates an object of class aop_cytoscape
}
\section{Slots}{

   \describe{
     \item{\code{name}:}{Object of class \code{"character"}, containing the
     name of the AOP.}
     \item{\code{nodes}:}{Object of class \code{"list"}, containing the list
     of nodes.}
     \item{\code{edges}:}{Object of class \code{"list"}, contianing the list of
     edges.}
   }
}
\author{
Lyle D. Burgoon
}

