% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bmr_Class.R
\docType{class}
\name{bmr-class}
\alias{bmr}
\alias{bmr-class}
\title{bmr class}
\description{
Creates an object of class bmr (bootstrap metaregression)
}
\section{Slots}{

   \describe{
     \item{\code{models}:}{Object of class \code{"list"}, containing the
     models.}
     \item{\code{fits}:}{Object of class \code{"list"}, containing the model
     fit predictions.}
     \item{\code{medians}:}{Object of class \code{"vector"}, containing the
     medians.}
     \item{\code{confidence_envelope}:}{Object of class \code{"data.frame"},
     containing the lower and upper confidence bounds.}
   }
}
\author{
Lyle D. Burgoon
}

