% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pod_analysis.R
\name{calculate_confidence_and_median}
\alias{calculate_confidence_and_median}
\title{Calculate the confidence envelope and the median for the bootstrap
metaregression concentration response models.}
\usage{
calculate_confidence_and_median(bootstrap_metaregression_obj)
}
\arguments{
\item{bootstrap_metaregression_obj}{the object that contains the bootstrap
metaregression models as a \code{bmr} object.}
}
\value{
bmr_obj a \code{bmr} object that holds all of the bootstrap
metaregression models produced.
}
\description{
Calculates the 95% confidence envelope and the median for
concentration-response data.
}
\details{
This is an internal function that calculates the 95% confidence envelope and
the median for concentration-response data. It is called by the
bootstrap_metaregression function.
}
\examples{
bmr_obj <- bootstrap_metaregression(oxybenzone, 15, 100)
}

