% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pod_analysis.R
\name{internal_model_fits}
\alias{internal_model_fits}
\title{An internal function for calculating interpolation values.}
\usage{
internal_model_fits(bmr_model, interval)
}
\arguments{
\item{bmr_model}{a model of class \code{lm} based on a bootstrap natural
splin metaregression of concentration-response data.}

\item{interval}{a \code{numeric} value that specifies how large the
interval should be between each value used for interpolation.}
}
\value{
temp_df a \code{data.frame} consisting of concentration and
activity columns that represent the interpolated model-based
response/activity values.
}
\description{
An internal function for calculating interpolation values.
}

