% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pod_analysis.R
\docType{data}
\name{oxybenzone}
\alias{oxybenzone}
\title{High Throughput Screening Data (Tox21) for Assessing the Estrogenicity of
Oxybenzone}
\format{A data frame with 58 rows and 4 variables}
\usage{
data(oxybenzone)
}
\description{
A dataset containing the concentration-response data for analyzing the
estrogenicity of oxybenzone from PubChem (Assay ID: 743075, Substance ID:
144209183, Chemical ID: 4632; Assay ID: 743079, Substance ID: 144203969
Chemical ID: 4632). 125 rows. 58 rows x 4 cols/variables.
}
\details{
AssayDataset DatasetReplicate Concentration  Activity

\itemize{
  \item AssayDataset. Coded value (1/2) that corresponds to an Assay ID
  \item DatasetReplicate. This is the "biological" replicate within an
  AssayDataset (1--3)
  \item Concentration. This is the concentration of the chemical in the assay (5.55694e-04--7.03500e+01)
  \item Activity. This is the assay activity. Percent response for these assays. (1.646933e-03--1.140800e+02)
}
}
\keyword{datasets}

