% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palette.R
\name{plot_palette}
\alias{plot_palette}
\title{Generate a palette function}
\usage{
plot_palette(
  name,
  type = c("discrete", "continuous"),
  reverse = FALSE,
  ensure_contrast = FALSE
)
}
\arguments{
\item{name}{Name of the palette}

\item{type}{"discrete" or "continuous"}

\item{reverse}{Reverse palette order?}
}
\value{
A function(n) returning n colours
}
\description{
Generate a palette function
}
\keyword{internal}
