% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pod_analysis.R
\name{pod_slope_analysis}
\alias{pod_slope_analysis}
\title{This is intended as an internal function to facilitate the identification
of the concentration where the asymptote is likely to end based on a slope
threshold, so that the mean of the upper confidence limit for the
asymptote can be calculated.}
\usage{
pod_slope_analysis(pod_slope_data, slope_threshold = 10)
}
\arguments{
\item{pod_slope_data}{the \code{data.frame} object that contains the
concentration and slope data.}

\item{slope_threshold}{the \code{numeric} object that sets the threshold
for the slope. This determines the upper bound on the concentration range
that is determined to be the asymptote. In other words, the asymptote is
defined as that region that has a slope less than the threshold at the lower
end of the concentration-response curve.}
}
\description{
An internal function for calculating the boundary of the lower asymptote.
}

