% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_custom.R
\name{scale_fill_custom}
\alias{scale_fill_custom}
\title{Custom fill scale from user-supplied palette}
\usage{
scale_fill_custom(palette, discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{A vector of colours (hex codes or names)}

\item{discrete}{Whether to treat the scale as discrete or continuous}

\item{reverse}{Whether to reverse the palette order}

\item{...}{Additional arguments passed to the scale}
}
\value{
A ggplot2 fill scale
}
\description{
Custom fill scale from user-supplied palette
}
