% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop_palette.R
\name{aop_palette}
\alias{aop_palette}
\title{Retrieve colours from an aop palette}
\usage{
aop_palette(name)
}
\arguments{
\item{name}{Palette name (e.g., "sunset")}
}
\value{
Character vector of hex colour codes
}
\description{
Retrieve colours from an aop palette
}
