% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aop_palettes.R
\docType{data}
\name{aop_palettes}
\alias{aop_palettes}
\title{Colour palettes for Anatomy of Plots}
\format{
An object of class \code{list} of length 3.
}
\usage{
aop_palettes
}
\description{
This object contains all named palettes used throughout the package.
Each palette is a named character vector of hex codes.
}
\keyword{internal}
