% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_plot_base.R
\name{scale_colour_plot}
\alias{scale_colour_plot}
\title{Colour scale for Anatomy of Plots palettes}
\usage{
scale_colour_plot(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::discrete_scale()} or \code{ggplot2::scale_fill_gradientn()}.}
}
\description{
Colour scale for Anatomy of Plots palettes
}
