% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_plot.R
\name{scale_fill_plot}
\alias{scale_fill_plot}
\title{Fill scale for Anatomy of Plots palettes}
\usage{
scale_fill_plot(palette = "sunset", discrete = TRUE, reverse = FALSE, ...)
}
\description{
Fill scale for Anatomy of Plots palettes
}
