% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_plot_base.R
\name{scale_fill_plot}
\alias{scale_fill_plot}
\title{Fill scale for Anatomy of Plots palettes}
\usage{
scale_fill_plot(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{ggplot2::discrete_scale()} or \code{ggplot2::scale_fill_gradientn()}.}
}
\description{
Fill scale for Anatomy of Plots palettes
}
