% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_palette_selection.R
\name{test_palette_selection}
\alias{test_palette_selection}
\title{Visualize palette and its most distinct subset}
\usage{
test_palette_selection(palette, k = 2, title = NULL, dark_mode = FALSE)
}
\arguments{
\item{palette}{A character vector of hex colour codes}

\item{k}{Number of colours to select}

\item{title}{Optional plot title}
}
\value{
A ggplot comparing full vs selected palette
}
\description{
Visualize palette and its most distinct subset
}
